fix(BukkitHistoryService): add getMessageIds method to retrieve cached message IDs

+++ surf-chat-bukkit/src/main/kotlin/dev/slne/surf/chat/bukkit/command/surfchat/SurfChatDeleteCommand.kt
@@ -4,6 +4,7 @@ import dev.jorel.commandapi.CommandAPICommand
 import dev.jorel.commandapi.kotlindsl.playerExecutor
 import dev.jorel.commandapi.kotlindsl.stringArgument
 import dev.slne.surf.chat.api.surfChatApi
+import dev.slne.surf.chat.bukkit.service.BukkitHistoryService
 import dev.slne.surf.chat.core.service.historyService
 import dev.slne.surf.surfapi.core.api.messages.adventure.buildText
 import java.util.*
@@ -14,6 +15,15 @@ class SurfChatDeleteCommand(commandName: String) : CommandAPICommand(commandName
         stringArgument("messageID")
         playerExecutor { player, args ->
             val messageID = args.getUnchecked<String>("messageID") ?: return@playerExecutor
+
+            if(historyService.getMessageIds().contains(UUID.fromString(messageID))) {
+                surfChatApi.sendText(player, buildText {
+                    error("Eine Nachricht mit der ID ")
+                    variableValue(messageID)
+                    error(" existiert nicht.")
+                })
+                return@playerExecutor
+            }
 
             if(historyService.deleteMessage(player.name, UUID.fromString(messageID))) {
                 surfChatApi.sendText(player, buildText {
